Allow debuggers to get the 'real' callsite.
The previous commit introduced a behaviour change: previously, the test
suite failed under the debugger, because a sub called from the debugger
would return a callsite somewhere in the debugger's guts. Changing to
caller_cx will properly skip those stack frames.

On the off-chance that someone might want that information, if callsite
is called in list context from package DB, return a 2-element list,
where the first element is the ordinary callsite and the second is the
real callsite of the current sub.